Add useDetailedItemNoState hook #5733
