+++ ui/app/src/components/LegacyFormDialog/EmbeddedLegacyContainer.tsx
@@ -48,8 +48,7 @@ import { translations } from './translations';
 import { useStyles } from './styles';
 import { hasEditAction } from '../../utils/content';
 import { nnou } from '../../utils/object';
-import useActiveSiteId from '../../hooks/useActiveSiteId';
-import { fetchDetailedItem } from '../../services/content';
+import { useDetailedItemNoState } from '../../hooks/useDetailedItemNoState';
 
 export const EmbeddedLegacyContainer = React.forwardRef(function EmbeddedLegacyEditor(
   props: LegacyFormDialogContainerProps,
@@ -80,15 +79,11 @@ export const EmbeddedLegacyContainer = React.forwardRef(function EmbeddedLegacyE
   const iframeRef = useRef(null);
   const dispatch = useDispatch();
   const [error, setError] = useState<ApiResponse>(null);
-  const [item, setItem] = useState(null);
-  const siteId = useActiveSiteId();
+  // When filename, path prop will still be the previous one, and useDetailedItem will try to re-fetch the
+  // non-existing item (old filename path), so we will only re-fetch when the actual path prop of the component
+  // changes (useDetailedItemNoState).
+  const item = useDetailedItemNoState(path);
   const availableActions = item?.availableActions;
-
-  // If filename changes, useDetailedItem will try to re-fetch a non-existing item (old path), so we will only re-fetch
-  // if path changes
-  useEffect(() => {
-    fetchDetailedItem(siteId, path).subscribe(setItem);
-  }, [siteId, path]);
 
   const src = useMemo(
     () =>
